Prompt/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
websockets==12.0
aiofiles==23.2.1